apps/worker/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
gitpython==3.1.41
tree-sitter==0.23.2
tree-sitter-javascript==0.23.1
tree-sitter-typescript==0.23.2
tree-sitter-python==0.23.5
google-generativeai==0.3.2
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.1
aiofiles==23.2.1
